Refactor ToCore
