+++ library/Core/src/Juvix/Core/Base/Types/Base.hs
@@ -127,6 +127,13 @@ type CoreShow ext primTy primVal = (
   , Show (XApp ext primTy primVal)
   , Show (XAnn ext primTy primVal)
   , Show (ElimX ext primTy primVal)
+  , Show (XPCon ext primTy primVal)
+  , Show (XPPair ext primTy primVal)
+  , Show (XPUnit ext primTy primVal)
+  , Show (XPVar ext primTy primVal)
+  , Show (XPDot ext primTy primVal)
+  , Show (XPPrim ext primTy primVal)
+  , Show (PatternX ext primTy primVal)
   )
 
 type CoreEq ext primTy primVal = (
@@ -147,4 +154,11 @@ type CoreEq ext primTy primVal = (
   , Eq (XApp ext primTy primVal)
   , Eq (XAnn ext primTy primVal)
   , Eq (ElimX ext primTy primVal)
+  , Eq (XPCon ext primTy primVal)
+  , Eq (XPPair ext primTy primVal)
+  , Eq (XPUnit ext primTy primVal)
+  , Eq (XPVar ext primTy primVal)
+  , Eq (XPDot ext primTy primVal)
+  , Eq (XPPrim ext primTy primVal)
+  , Eq (PatternX ext primTy primVal)
   )
